pooling: merge src/ocl/pooling_ocl.cpp with src/pooling.cpp
Now that these kernels are ported to HIP, src/ocl/pooling_ocl.cpp
is not really a sensible place anymore. The code from that file is
now merged with the rest of the PoolingDescriptor code.
